backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kanban-backend"
version = "0.1.0"
description = "FastAPI backend for realtime Kanban MVP"
readme = "README.md"
authors = [{ name = "Kanban Team" }]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.1",
  "python-socketio[asgi]>=5.11.0",
  "SQLAlchemy>=2.0.29",
  "asyncpg>=0.29.0",
  "psycopg[binary]>=3.1.18",
  "alembic>=1.13.1",
  "pydantic[email]>=2.6.3",
  "pydantic-settings>=2.2.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "redis>=5.0.1",
  "fastapi-limiter>=0.1.6",
  "structlog>=24.1.0",
  "loguru>=0.7.2",
  "argon2-cffi>=23.1.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.3.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.5",
  "pytest-cov>=4.1.0",
  "httpx>=0.27.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
